+++ src/compas_model/viewer/modelobject.py
@@ -1,0 +1,6 @@
+from compas_model.scene.modelobject import ModelObject as BaseModelObject
+from compas_viewer.scene import ViewerSceneObject
+
+
+class ModelObject(ViewerSceneObject, BaseModelObject):
+    pass